Fix: 활동 인증 수정, 로컬 환경 테스트 종료

See merge request s10-bigdata-recom-sub2/S10P22A707!151

Former-commit-id: bb316baeb39eaae6bc43a2edb29412094a2d2be0

+++ backend/src/main/java/com/tfeo/backend/domain/member/controller/WishController.java
@@ -1,6 +1,7 @@
 package com.tfeo.backend.domain.member.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.tfeo.backend.common.model.dto.SuccessResponse;
 import com.tfeo.backend.common.service.AuthenticationService;
 import com.tfeo.backend.domain.member.model.dto.FindWishListDto;
+import com.tfeo.backend.domain.member.model.entity.Member;
 import com.tfeo.backend.domain.member.service.WishService;
 
 import lombok.RequiredArgsConstructor;
@@ -29,13 +31,12 @@ public class WishController {
 
 	@PostMapping("/{homeNo}")
 	public ResponseEntity<?> wishAdd(@PathVariable("homeNo") Long homeNo, HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L;
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		wishService.addWish(memberNo, homeNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
@@ -46,13 +47,12 @@ public class WishController {
 
 	@GetMapping("")
 	public ResponseEntity<?> wishList(HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L;
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		List<FindWishListDto> findWishListDtoList = wishService.findWishList(memberNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
@@ -63,13 +63,12 @@ public class WishController {
 
 	@DeleteMapping("/{homeNo}")
 	public ResponseEntity<?> wishRemove(@PathVariable("homeNo") Long homeNo, HttpServletRequest request) {
-		// Optional<Member> memberOptional = authenticationService.getMember(request);
-		// if (!memberOptional.isPresent()) {
-		// 	return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
-		// }
-		// Member member = memberOptional.get();
-		// Long memberNo = member.getMemberNo();
-		Long memberNo = 1L;
+		Optional<Member> memberOptional = authenticationService.getMember(request);
+		if (!memberOptional.isPresent()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("가입된 사용자 데이터를 찾을 수 없습니다.");
+		}
+		Member member = memberOptional.get();
+		Long memberNo = member.getMemberNo();
 		wishService.removeWish(memberNo, homeNo);
 		SuccessResponse successResponse = SuccessResponse.builder()
 			.status(HttpStatus.OK)
